requirements.txt
selenium==4.0.0
appium-python-client==3.2.1
pillow==10.4.0
apscheduler==3.8.0
pytesseract==0.3.8
webdriver-manager==4.0.2
